% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{dbConnection}
\alias{dbConnection}
\title{Baut eine Verbindung zur FVA-Fotofallendatenbank auf (PostgreSQL).}
\usage{
dbConnection(type)
}
\arguments{
\item{type}{Bestimmt welche Art Verbindung hergestellt werden soll, ob nur Lese- oder auch Schreibrechte: type = c("read","write").}
}
\value{
Formal class PostgreSQLConnection
}
\description{
Baut eine Verbindung zur FVA-Fotofallendatenbank auf (PostgreSQL).
}
\examples{

}
