% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectionUpload.R
\name{dbConnectionUpload}
\alias{dbConnectionUpload}
\title{Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu überprüfen.}
\usage{
dbConnectionUpload(psw)
}
\arguments{
\item{psw}{Passwort für den Zugriff auf die FVA-Fotofallendatenbank mit Schreibrechten.}
}
\value{
Formal class PostgreSQLConnection
}
\description{
Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu überprüfen.
}
\examples{

}
